Fix Fmt enum AVR truncation and HtmlFmt's missing fmtStart/fmtStop passthrough
enums.h: Fmt was int-backed but needs 22 independent bits (up to
Footer=1<<21) — int is only 16 bits on AVR, so Option/Selected/Choice/
Dropdown/Enabled silently collided with earlier tags there. Surfaced by
building XmlFmt for a true-16-bit-int AVR target (Uno) for the first time.
Widened the enum and its |/&/^ operator templates to long — a 3+-way OR
chain needs the whole chain in long, not just the final result, since it
associates left-to-right through the same operators.

xmlFmt.h: attr_tags/indent_tags/block_tags stored that OR'd result into
const int, truncating it right back — widened to long too.

htmlFmt.h: missing `using Base::fmtStart; using Base::fmtStop;` (present
in textFmt.h/gfxFmt.h, the other per-tag-overload-style formats) meant its
own overloads shadowed the base's default no-op for every tag they don't
handle, not just override the ones they do — "no matching function" for
Fmt::Index, which IndexPrinter always calls. HtmlFmt has zero other call
sites anywhere in the repo; this was never actually exercised before.

+++ include/oneMenu/menu/fmt/xmlFmt.h
@@ -83,9 +83,13 @@ namespace oneMenu  {
       // the others) — still fit attr_tags cleanly since they only ever need
       // "is there an open tag to attach to" (always true: Choice fires right
       // after <menu>/<item> opens, Dropdown right after <item> opens).
-      static constexpr const int attr_tags   = Fmt::NavCursor|Fmt::Index|Fmt::EditMode|Fmt::Accel|Fmt::Choice|Fmt::Dropdown|Fmt::Enabled;
-      static constexpr const int indent_tags = Fmt::View|Fmt::Menu|Fmt::Body|Fmt::Title|Fmt::Item;
-      static constexpr const int block_tags  = Fmt::View|Fmt::Menu|Fmt::Body|Fmt::Title|Fmt::Item;
+      // long, not int: Choice/Dropdown/Enabled live above bit 15 (enums.h) —
+      // int is only 16 bits on AVR, so storing this OR'd mask into an int
+      // silently dropped those bits there (found alongside the enums.h fix
+      // this comment references, same root cause).
+      static constexpr const long attr_tags   = Fmt::NavCursor|Fmt::Index|Fmt::EditMode|Fmt::Accel|Fmt::Choice|Fmt::Dropdown|Fmt::Enabled;
+      static constexpr const long indent_tags = Fmt::View|Fmt::Menu|Fmt::Body|Fmt::Title|Fmt::Item;
+      static constexpr const long block_tags  = Fmt::View|Fmt::Menu|Fmt::Body|Fmt::Title|Fmt::Item;
 
       void putPath(const Path& p, Depth s, Depth l) {
         Depth end = (s+l < p.len) ? s+l : p.len-1;
